+++ src/main/java/com/project/popupmarket/controller/invitation/InvitationController.java
@@ -1,8 +1,7 @@
-package com.project.popupmarket.controller;
+package com.project.popupmarket.controller.invitation;
 
 import com.project.popupmarket.dto.invitation.InvitationTO;
-import com.project.popupmarket.dto.Resp;
-import com.project.popupmarket.service.InvitationService;
+import com.project.popupmarket.service.invitation.InvitationService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,14 +16,14 @@ public class InvitationController {
     }
 
     @PostMapping("/invitation")
-    public ResponseEntity<Resp> addInvitation(@RequestBody InvitationTO invitation) {
+    public ResponseEntity<String> addInvitation(@RequestBody InvitationTO invitation) {
 
         boolean flag = invitationService.addInvitation(invitation);
 
         if (flag) {
-            return ResponseEntity.ok(new Resp(200, "success"));
+            return ResponseEntity.ok("success");
         } else {
-            return ResponseEntity.status(500).body(new Resp(500, "error"));
+            return ResponseEntity.status(500).body("error");
         }
     }
 
@@ -43,14 +42,14 @@ public class InvitationController {
 //    }
 
     @DeleteMapping("/invitation")
-    public ResponseEntity<Resp> removeInvitation(@RequestBody InvitationTO invitation) {
+    public ResponseEntity<String> removeInvitation(@RequestBody InvitationTO invitation) {
 
         boolean flag = invitationService.removeInvitation(invitation);
 
         if (flag) {
-            return ResponseEntity.ok(new Resp(200, "success"));
+            return ResponseEntity.ok("success");
         } else {
-            return ResponseEntity.status(500).body(new Resp(500, "error"));
+            return ResponseEntity.status(500).body("error");
         }
     }
 }
