--- src/main/java/com/project/popupmarket/dto/Resp.java
@@ -1,13 +1,0 @@
-package com.project.popupmarket.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Resp {
-    private int status;
-    private String message;
-}
